\name{cuberoot}
\alias{cuberoot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculates the cubic root
}
\description{
cuberoot(x) computes the cubic root of x, \eqn{\sqrt[3]{x}}.
}
\usage{
cuberoot(x)
}
\arguments{
  \item{x}{
a numeric or complex vector or array.
}
}
\value{
the cube root of a number.
}
\author{
Diego Gallardo
}

\examples{
cuberoot(-27)
cuberoot(0)
cuberoot(64)
}

