\name{denmark}
\alias{denmark}
\docType{data}
\encoding{latin1}
\title{Data set for Denmark, Johansen and Juselius (1990)}
\description{
  This data set contains the series used by S. Johansen and K. Juselius
  for estimating a money demand function of Denmark.
}
\usage{data(denmark)}
\format{
  A data frame with 55 observations on the following 6 variables.

    \tabular{rl}{
    \code{period} \tab Time index from 1974:Q1 until 1987:Q3. \cr 
    \code{LRM} \tab Logarithm of real money, M2. \cr
    \code{LRY} \tab Logarithm of real income. \cr
    \code{LPY} \tab Logarithm of price deflator. \cr
    \code{IBO} \tab Bond rate. \cr
    \code{IDE} \tab Bank deposit rate. \cr
    }
}
\source{
  Johansen, S. and Juselius, K. (1990), Maximum Likelihood Estimation and
  Inference on Cointegration -- with Applications to the Demand for
  Money, \emph{Oxford Bulletin of Economics and Statistics}, \bold{52,
    2}, 169--210.
  
}

\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Johansen Juselius Denmark}
