% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_explore_tab_culvert_count_by_WRIA_histogram}
\alias{figure_explore_tab_culvert_count_by_WRIA_histogram}
\title{figure explore tab culvert count by WRIA histogram}
\usage{
figure_explore_tab_culvert_count_by_WRIA_histogram(sfC, sfW)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{sfW}{simple features data frame of wria data}
}
\value{
ggplot object of culvert count by WRIA
}
\description{
figure explore tab culvert count by WRIA histogram
}
