% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_explore_tab_histogram}
\alias{figure_explore_tab_histogram}
\title{figure explore tab histogram}
\usage{
figure_explore_tab_histogram(sfC, r)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{r}{reactive values object with app input values}
}
\value{
ggplot object of culvert data histogram
}
\description{
figure explore tab histogram
}
