% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_explore_tab_scatterplot}
\alias{figure_explore_tab_scatterplot}
\title{figure explore tab scatterplot}
\usage{
figure_explore_tab_scatterplot(sfC, r)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{r}{reactive values object with app input values}
}
\value{
ggplot object of culvert data scatterplot
}
\description{
figure explore tab scatterplot
}
