% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_histogram}
\alias{figure_histogram}
\title{figure explore tab histogram}
\usage{
figure_histogram(
  points,
  x_axis_variable,
  y_axis_variable,
  color_variable,
  histogram_variable,
  histogram_nbins,
  highlight,
  barrier_ids,
  plot_xmin,
  plot_xmax,
  plot_ymin,
  plot_ymax
)
}
\arguments{
\item{points}{simple features data frame of culvert data}
}
\value{
ggplot object of culvert data histogram
}
\description{
figure explore tab histogram
}
