% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_scatterplot}
\alias{figure_scatterplot}
\title{Figure explore tab scatterplot}
\usage{
figure_scatterplot(
  points,
  x_axis_variable,
  y_axis_variable,
  color_variable,
  x_jitter,
  y_jitter,
  highlight,
  barrier_ids,
  plot_xmin,
  plot_xmax,
  plot_ymin,
  plot_ymax
)
}
\arguments{
\item{points}{simple features data frame of culvert data}

\item{x_axis_variable}{}

\item{y_axis_variable}{}

\item{color_variable}{}

\item{x_jitter}{}

\item{y_jitter}{}

\item{highlight}{}

\item{barrier_ids}{}

\item{plot_xmin}{}

\item{plot_xmax}{}

\item{plot_ymin}{}
}
\value{
ggplot object of culvert data scatterplot
}
\description{
Figure explore tab scatterplot
}
