% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{filter_and_format_culverts_for_explore_tab_scatterplot}
\alias{filter_and_format_culverts_for_explore_tab_scatterplot}
\title{filter and format culverts for explore tab scatterplot}
\usage{
filter_and_format_culverts_for_explore_tab_scatterplot(sfC, r)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{r}{reactive values object with app input values}
}
\value{
data frame of culvert data formatted to make scatterplot in ggplot
}
\description{
filter and format culverts for explore tab scatterplot
}
