% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{filter_and_format_culverts_for_histogram}
\alias{filter_and_format_culverts_for_histogram}
\title{filter and format culverts for histogram}
\usage{
filter_and_format_culverts_for_histogram(
  points,
  area_sel,
  owner_sel,
  color_variable,
  histogram_variable
)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{area_sel}{A vector of WRIA ID numbers of interest.}

\item{owner_sel}{A vector of owner ID numbers of interest.}
}
\value{
A data frame of point data formatted to make histogram in ggplot.
}
\description{
filter and format culverts for histogram
}
