% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{filter_and_format_culverts_for_scatterplot}
\alias{filter_and_format_culverts_for_scatterplot}
\title{filter and format culverts for explore tab scatterplot}
\usage{
filter_and_format_culverts_for_scatterplot(
  sfC,
  area_sel,
  owner_sel,
  x_axis_variable,
  y_axis_variable,
  color_variable
)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{r}{reactive values object with app input values}
}
\value{
data frame of culvert data formatted to make scatterplot in ggplot
}
\description{
filter and format culverts for explore tab scatterplot
}
