% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{get_plot_click_site_id}
\alias{get_plot_click_site_id}
\title{get plot click site id}
\usage{
get_plot_click_site_id(
  points,
  owner_sel,
  area_sel,
  remove_bad_match,
  x_axis_variable,
  y_axis_variable,
  plotClickX,
  plotClickY
)
}
\arguments{
\item{points}{points}

\item{owner_sel}{owner selection}

\item{area_sel}{wria selection}

\item{remove_bad_match}{indicator for removing a bad match}

\item{x_axis_variable}{variable on the x axis}

\item{y_axis_variable}{variable on the y axis}

\item{plotClickX}{x coordinate from plot click event}

\item{plotClickY}{y coordinate from plot click event}
}
\value{
string value of culvert site id closest to plot click coordinates or empty string if beyond maximum distance
}
\description{
get plot click site id
}
