% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom.R
\name{get_points_sel_hab_custom}
\alias{get_points_sel_hab_custom}
\title{Convert projects in custom portfolio adding habitat to TRUE/FALSE vector}
\usage{
get_points_sel_hab_custom(points, prtf_cust, barrier_idp, E)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{prtf_cust}{A vector of point IDs for selected points.}

\item{barrier_idp}{A vector of planned culvert IDs.}

\item{E}{A full connectivity matrix.}
}
\value{
A simple features point data frame containing culvert location and attributes for selected points.
}
\description{
Convert projects in custom portfolio adding habitat to TRUE/FALSE vector
}
