% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{get_pretty_variable_name}
\alias{get_pretty_variable_name}
\title{get pretty variable name}
\usage{
get_pretty_variable_name(varName)
}
\arguments{
\item{varName}{string of variable name}
}
\value{
string value of variable name to use in chart axis and legend labels
}
\description{

}
\details{

}
