% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Tables.R
\name{get_project_list}
\alias{get_project_list}
\title{Display a list of suggested projects}
\usage{
get_project_list(points, points_sel)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{points_sel}{A simple features point data frame containing selected culvert locations and attributes.}
}
\value{
A dataframe listing suggested projects
}
\description{
Display a list of suggested projects
}
