% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Tables.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Summarize variables for select culverts}
\usage{
get_summary_table(points, points_sel, barrier_idp)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{points_sel}{A simple features point data frame containing selected culvert locations and attributes.}

\item{barrier_idp}{A vector of planned culvert IDs}
}
\value{
A summary table
}
\description{
Summarize variables for select culverts
}
