% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{get_wria_bounding_box}
\alias{get_wria_bounding_box}
\title{Get a bounding box for selected WRIAS}
\usage{
get_wria_bounding_box(area_sel)
}
\arguments{
\item{area_sel}{A vector of WRIA numbers of interest.}
}
\value{
A numeric vector of the bounding box for the provided WRIA simple features object.
}
\description{
Get a bounding box for selected WRIAS
}
