% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom.R
\name{map_leaflet_custom}
\alias{map_leaflet_custom}
\title{Map the custom solution}
\usage{
map_leaflet_custom(leaf_proxy, points, lines, prtf_cust, E, marginal_line_ids)
}
\arguments{
\item{leaf_proxy}{A basemap.}

\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{lines}{A simple features data frame with linestring geometries.}

\item{prtf_cust}{A vector of point IDs for selected points.}

\item{E}{A full connectivity matrix.}

\item{marginal_line_ids}{A vector of line IDs for all lines marginally upstream of each point.}
}
\value{
A leaflet map.
}
\description{
Map the custom solution
}
