% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Suggest.R
\name{map_leaflet_opt}
\alias{map_leaflet_opt}
\title{Map the optimization solution}
\usage{
map_leaflet_opt(
  leaf_proxy,
  points,
  lines,
  dslines,
  soln,
  marginal_line_ids,
  downstream_line_ids,
  wria_sel,
  huc_sel
)
}
\arguments{
\item{leaf_proxy}{A basemap.}

\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{lines}{A simple features data frame with linestring geometries upstream of culvs}

\item{dslines}{A simple features data frame with linestring geometries downstream of culvs}

\item{soln}{A logical vector of TRUE/FALSE values that is the output from solve_opt(.}

\item{marginal_line_ids}{A vector of line IDs for all lines marginally upstream of each point.}

\item{downstream_line_ids}{A vector of line IDs for all lines marginally upstream of each point.}

\item{area_sel}{A vector of WRIA ID numbers of interest.}

\item{subarea_sel}{A vector of WRIA ID numbers of interest.}
}
\value{
A leaflet map.
}
\description{
Map the optimization solution
}
