% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{myGeneric}
\alias{myGeneric}
\title{A Generic}
\usage{
myGeneric(x, y, ...)
}
\arguments{
\item{x}{Describe x.}

\item{y}{Describe y.}

\item{...}{Describe ... .}
}
\description{
This generic always takes x,y, other methods might use ...
}
