\name{Raotbl1} 
\alias{Raotbl1}
\docType{data}
\encoding{latin1}
\title{Data set used by Dickey, Jansen and Thornton (1994)}
\description{
  This data set contains the time series used by David A. Dickey, Dennis
  W. Jansen and Daniel L. Thornton in their article: ``A Primer on
  Cointegrating with an Application to Money and Income''. 
}
\usage{data(Raotbl1)}
\format{
  A data frame with quarterly oberservations (ts objects) starting in
  1953:1 until 1988:4 for the following 4 variables (all transformed to
  natural logarithms.
  
    \tabular{rl}{
    \code{k} \tab Ratio of currency to total checkable deposits. \cr
    \code{ksa} \tab seasonally adjusted series of \code{k}. \cr
    \code{r3m} \tab Nominal 3 month T-Bill rate. \cr
    \code{r10y} \tab Nominal yield on 10-year Government securities. \cr
    \code{rgnp} \tab Real GNP.
    }
}
\source{
  Dickey, David A., Dennis W. Jansen and Daniel L. Thornton (1994), A
  Primer on Cointegration with an Application to Money and Income, in:
  Cointegration for the Applied Economist, ed. B. Bhaskara Rao, chapter
  2, Data Appendix, Table D.1.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Dickey Jansen Thornton}
