% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{remove_map_points}
\alias{remove_map_points}
\title{Remove map points}
\usage{
remove_map_points(leaf_proxy)
}
\arguments{
\item{leaf_proxy}{An leaflet map with culverts to be cleared.}
}
\value{
none
}
\description{
Remove map points
}
