% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{replace_owner_code_with_name}
\alias{replace_owner_code_with_name}
\title{replace owner code with name}
\usage{
replace_owner_code_with_name(points)
}
\arguments{
\item{points}{simple features data frame of culvert data}
}
\value{
data frame of culverts with owner name instead of owner type code
}
\description{
replace owner code with name
}
