% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{replace_WRIA_number_with_name}
\alias{replace_WRIA_number_with_name}
\title{replace WRIA number with name}
\usage{
replace_WRIA_number_with_name(sfC, sfW)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{sfW}{simple features data frame of wria data}
}
\value{
data frame of culverts with wria name instead of wria number
}
\description{
replace WRIA number with name
}
