% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{reset_map}
\alias{reset_map}
\title{Reset the map}
\usage{
reset_map(leaf_proxy)
}
\arguments{
\item{leaf_proxy}{An leaflet map with groups to be cleared.}
}
\value{
none
}
\description{
Reset the map
}
