% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Suggest.R
\name{solve_opt}
\alias{solve_opt}
\title{Solve the optimization problem}
\usage{
solve_opt(points, budget, D, wria_sel, owner_sel)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{budget}{A number specifying the budget constraint.}

\item{D}{A connectivity matrix.}

\item{owner_sel}{A vector of owner ID numbers of interest.}

\item{area_sel}{A vector of WRIA ID numbers of interest.}
}
\value{
A logical vector of TRUE/FALSE values.
}
\description{
Solve the optimization problem
}
