% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_culvert_markers}
\alias{update_map_culvert_markers}
\title{update map culvert markers}
\usage{
update_map_culvert_markers(
  leaf_proxy,
  area_sel,
  owner_sel,
  color_variable,
  highlight,
  barrier_ids
)
}
\arguments{
\item{leaf_proxy}{leaflet proxy reference}

\item{r}{reactive values object with app input values}
}
\value{
none
}
\description{
update map culvert markers
}
