% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_selected_polygons}
\alias{update_map_selected_polygons}
\title{update map selected wria polygons}
\usage{
update_map_selected_polygons(leaf_proxy, area_sel, subarea_sel)
}
\arguments{
\item{leaf_proxy}{A leaflet proxy reference.}

\item{area_sel}{A vector of WRIA ID numbers of interest.}

\item{subarea_sel}{A vector of HUC 12 numbers of interest.}
}
\value{
none
}
\description{
update map selected wria polygons
}
