% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_WRIA_labels}
\alias{update_map_WRIA_labels}
\title{update map WRIA labels}
\usage{
update_map_WRIA_labels(leafProxy, zoomLevel, r)
}
\arguments{
\item{leafProxy}{leaflet proxy reference}

\item{zoomLevel}{zoom level for leaflet map}

\item{r}{reactive values object with app input values}
}
\value{
none
}
\description{
update map WRIA labels
}
