% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_scatterplot}
\alias{figure_scatterplot}
\title{figure explore tab scatterplot}
\usage{
figure_scatterplot(
  sfC,
  x_axis_variable,
  y_axis_variable,
  color_variable,
  x_jitter,
  y_jitter,
  highlight,
  barrier_ids,
  plot_xmin,
  plot_xmax,
  plot_ymin,
  plot_ymax
)
}
\arguments{
\item{sfC}{simple features data frame of culvert data}

\item{r}{reactive values object with app input values}
}
\value{
ggplot object of culvert data scatterplot
}
\description{
figure explore tab scatterplot
}
