% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{figure_scatterplot}
\alias{figure_scatterplot}
\title{Figure explore tab scatterplot}
\usage{
figure_scatterplot(
  points,
  x_axis_variable,
  y_axis_variable,
  color_variable,
  x_jitter,
  y_jitter,
  highlight,
  barrier_ids,
  plot_xmin,
  plot_xmax,
  plot_ymin,
  plot_ymax
)
}
\arguments{
\item{points}{simple features data frame of culvert data}

\item{x_axis_variable}{A variable to go on the x axis.}

\item{y_axis_variable}{A variable to go on the y axis.}

\item{color_variable}{A variable defining color palate.}

\item{x_jitter}{Level of jitter in the x dimension.}

\item{y_jitter}{Level of jitter in the y dimention.}

\item{highlight}{Whether to highlight some points.}

\item{barrier_ids}{IDs of points to highlight.}

\item{plot_xmin}{The minimum x value of the plot.}

\item{plot_xmax}{The maximum x value of the plot.}

\item{plot_ymin}{The minimum y value of the plot.}

\item{plot_ymax}{The maximum y value of the plot.}
}
\value{
ggplot object of culvert data scatterplot
}
\description{
Figure explore tab scatterplot
}
