% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Tables.R
\name{get_plan_list}
\alias{get_plan_list}
\title{Get the list of culverts selected for the plan}
\usage{
get_plan_list(
  points,
  points_sel,
  barrier_idp,
  remove_bad_match,
  cost = 1,
  mean_design_cost = NULL,
  mean_construction_cost = NULL
)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{points_sel}{A simple features point data frame containing selected culvert locations and attributes.}

\item{barrier_idp}{A vector of planned culvert IDs}
}
\value{
A summary table
}
\description{
Get the list of culverts selected for the plan
}
