% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom.R
\name{get_points_sel_custom}
\alias{get_points_sel_custom}
\title{Convert projects in custom portfolio to TRUE/FALSE vector}
\usage{
get_points_sel_custom(points, prtf_cust)
}
\arguments{
\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{prtf_cust}{A vector of point IDs for selected points.}
}
\description{
Convert projects in custom portfolio to TRUE/FALSE vector
}
