% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{get_wria_bounding_box}
\alias{get_wria_bounding_box}
\title{get bounding box for selected wrias}
\usage{
get_wria_bounding_box(cWRIAs)
}
\arguments{
\item{cWRIAs}{}
}
\value{
A numeric vector of the bounding box for the provided WRIA simple features
}
\description{
get bounding box for selected wrias
}
