% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Custom.R
\name{map_leaflet_custom}
\alias{map_leaflet_custom}
\title{Map the custom solution}
\usage{
map_leaflet_custom(
  leaf_proxy,
  points,
  lines,
  dslines,
  prtf_cust,
  E,
  marginal_line_ids,
  downstream_line_ids
)
}
\arguments{
\item{leaf_proxy}{A basemap.}

\item{points}{A simple features point data frame containing culvert locations and attributes.}

\item{lines}{A simple features data frame with linestring geometries.}

\item{dslines}{A simple features data frame with linestring geometries downstream of culvs}

\item{prtf_cust}{A vector of point IDs for selected points.}

\item{E}{A full connectivity matrix.}

\item{marginal_line_ids}{A vector of line IDs for all lines marginally upstream of each point.}

\item{downstream_line_ids}{A vector of line IDs for all lines downstream of each point.}
}
\value{
A leaflet map.
}
\description{
Map the custom solution
}
