% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_culvert_markers}
\alias{update_map_culvert_markers}
\title{update map culvert markers}
\usage{
update_map_culvert_markers(leafProxy, r)
}
\arguments{
\item{leafProxy}{leaflet proxy reference}

\item{r}{reactive values object with app input values}
}
\value{
none
}
\description{
update map culvert markers
}
