% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_culvert_markers}
\alias{update_map_culvert_markers}
\title{update map culvert markers}
\usage{
update_map_culvert_markers(
  leaf_proxy,
  area_sel,
  subarea_sel,
  owner_sel,
  color_variable,
  highlight,
  barrier_ids
)
}
\arguments{
\item{leaf_proxy}{leaflet proxy reference}

\item{area_sel}{A vector of WRIA ID numbers of interest.}

\item{subarea_sel}{A vector of WRIA ID numbers of interest.}

\item{owner_sel}{A vector of owner ID numbers of interest.}

\item{color_variable}{An attribute in the points data that defines point color.}

\item{highlight}{Set to NULL if there are no points to highlight.}

\item{barrier_ids}{A vector of point ids to highlight.}
}
\value{
none
}
\description{
update map culvert markers
}
