% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_selected_WRIA_polygons}
\alias{update_map_selected_WRIA_polygons}
\title{update map selected wria polygons}
\usage{
update_map_selected_WRIA_polygons(leafProxy, r)
}
\arguments{
\item{leafProxy}{leaflet proxy reference}

\item{r}{reactive values object with app input values}
}
\value{
none
}
\description{
update map selected wria polygons
}
