% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Figures.R
\name{update_map_WRIA_labels}
\alias{update_map_WRIA_labels}
\title{update map WRIA labels}
\usage{
update_map_WRIA_labels(leaf_proxy, zoom_level, area_sel, owner_sel)
}
\arguments{
\item{leaf_proxy}{A leaflet proxy reference.}

\item{zoom_level}{A zoom level for tge leaflet map.}

\item{area_sel}{A vector of WRIA ID numbers of interest.}

\item{owner_sel}{A vector of owner ID numbers of interest.}
}
\value{
none
}
\description{
update map WRIA labels
}
